# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /points/adjust:
    post:
      operationId: adjustPoints
      tags: [points]
      summary: Adjust points for the cousin (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustPointsBody"
      responses:
        "200":
          description: Points adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointsBalance"
  /redemptions:
    get:
      operationId: listRedemptions
      tags: [redemptions]
      summary: List redemption requests (admin sees all, cousin sees own)
      responses:
        "200":
          description: List of redemption requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedemptionList"
    post:
      operationId: createRedemption
      tags: [redemptions]
      summary: Create a redemption request (cousin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRedemptionBody"
      responses:
        "201":
          description: Redemption request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redemption"
  /redemptions/{id}/review:
    post:
      operationId: reviewRedemption
      tags: [redemptions]
      summary: Accept or deny a redemption request (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRedemptionBody"
      responses:
        "200":
          description: Redemption reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redemption"
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /admin/set-cousin:
    post:
      operationId: setCousinId
      tags: [points]
      summary: Set the cousin's user ID (admin only, first time setup)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCousinBody"
      responses:
        "200":
          description: Cousin set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CousinInfo"
...
components:
...
  schemas:
...
    PointsBalance:
      type: object
      required: [userId, points]
      properties:
        userId:
          type: string
        points:
          type: integer
...
    AdjustPointsBody:
      type: object
      required: [amount, reason]
      properties:
        amount:
          type: integer
        reason:
          type: string
    Redemption:
      type: object
      required: [id, userId, robuxAmount, pointsCost, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        robuxAmount:
          type: integer
        pointsCost:
          type: integer
        status:
          type: string
          enum: [pending, accepted, denied]
        note:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        reviewedAt:
          type: [string, 'null']
          format: date-time
    RedemptionList:
      type: object
      required: [redemptions]
      properties:
        redemptions:
          type: array
          items:
            $ref: "#/components/schemas/Redemption"
    CreateRedemptionBody:
      type: object
      required: [robuxAmount]
      properties:
        robuxAmount:
          type: integer
          minimum: 1
    ReviewRedemptionBody:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [accept, deny]
        note:
          type: [string, 'null']
    Notification:
      type: object
      required: [id, userId, message, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    CousinInfo:
      type: object
      required: [cousinId]
      properties:
        cousinId:
          type: [string, 'null']
        adminId:
          type: [string, 'null']
...
    SetCousinBody:
      type: object
      required: [cousinId]
      properties:
        cousinId:
          type: string