# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get the currently authenticated user
 */
export const GetCurrentAuthUserHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — Bearer <sid>."),
});

export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().email().nullable(),
      firstName: zod.string().nullable(),
      lastName: zod.string().nullable(),
      profileImageUrl: zod.string().nullable(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Start the browser OIDC login flow
 */
export const BeginBrowserLoginQueryParams = zod.object({
  returnTo: zod.coerce.string().optional(),
});

/**
 * @summary Complete the browser OIDC login flow
 */
export const HandleBrowserLoginCallbackQueryParams = zod.object({
  code: zod.coerce.string().optional(),
  state: zod.coerce.string().optional(),
  iss: zod.coerce.string().url().optional(),
});

/**
 * @summary Clear the session and begin OIDC logout
 */
export const LogoutBrowserSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — Bearer <sid>."),
});

/**
 * @summary Exchange a mobile OIDC code for a session token
 */

export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string().min(1),
  code_verifier: zod.string().min(1),
  redirect_uri: zod.string().url().min(1),
  state: zod.string().min(1),
  nonce: zod.string().min(1).optional(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Delete a mobile session token
 */
export const LogoutMobileSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — Bearer <sid>."),
});

export const LogoutMobileSessionResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current user's points balance
 */
export const GetMyPointsResponse = zod.object({
  userId: zod.string(),
  points: zod.number(),
});

/**
 * @summary Get points history for current user
 */
export const GetPointsHistoryResponse = zod.object({
  entries: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.string(),
      amount: zod.number(),
      reason: zod.string(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Adjust points for the cousin (admin only)
 */
export const AdjustPointsBody = zod.object({
  amount: zod.number(),
  reason: zod.string(),
});

export const AdjustPointsResponse = zod.object({
  userId: zod.string(),
  points: zod.number(),
});

/**
 * @summary List redemption requests (admin sees all, cousin sees own)
 */
export const ListRedemptionsResponse = zod.object({
  redemptions: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.string(),
      robuxAmount: zod.number(),
      pointsCost: zod.number(),
      status: zod.enum(["pending", "accepted", "denied"]),
      note: zod.string().nullish(),
      createdAt: zod.date(),
      reviewedAt: zod.date().nullish(),
    }),
  ),
});

/**
 * @summary Create a redemption request (cousin only)
 */

export const CreateRedemptionBody = zod.object({
  robuxAmount: zod.number().min(1),
});

/**
 * @summary Accept or deny a redemption request (admin only)
 */
export const ReviewRedemptionParams = zod.object({
  id: zod.coerce.number(),
});

export const ReviewRedemptionBody = zod.object({
  action: zod.enum(["accept", "deny"]),
  note: zod.string().nullish(),
});

export const ReviewRedemptionResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  robuxAmount: zod.number(),
  pointsCost: zod.number(),
  status: zod.enum(["pending", "accepted", "denied"]),
  note: zod.string().nullish(),
  createdAt: zod.date(),
  reviewedAt: zod.date().nullish(),
});

/**
 * @summary Get notifications for current user
 */
export const ListNotificationsResponse = zod.object({
  notifications: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.string(),
      message: zod.string(),
      read: zod.boolean(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Mark a notification as read
 */
export const MarkNotificationReadParams = zod.object({
  id: zod.coerce.number(),
});

export const MarkNotificationReadResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  message: zod.string(),
  read: zod.boolean(),
  createdAt: zod.date(),
});

/**
 * @summary Get the cousin's user ID (admin only)
 */
export const GetCousinIdResponse = zod.object({
  cousinId: zod.string().nullable(),
});

/**
 * @summary Set the cousin's user ID (admin only, first time setup)
 */
export const SetCousinIdBody = zod.object({
  cousinId: zod.string(),
});

export const SetCousinIdResponse = zod.object({
  cousinId: zod.string().nullable(),
});