# lib/api-zod/src/generated/types/notification.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Notification {
  id: number;
  userId: string;
  message: string;
  read: boolean;
  createdAt: Date;
}