# lib/api-zod/src/generated/types/redemption.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RedemptionStatus } from "./redemptionStatus";

export interface Redemption {
  id: number;
  userId: string;
  robuxAmount: number;
  pointsCost: number;
  status: RedemptionStatus;
  /** @nullable */
  note?: string | null;
  createdAt: Date;
  /** @nullable */
  reviewedAt?: Date | null;
}