# lib/api-zod/src/generated/types/redemptionList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Redemption } from "./redemption";

export interface RedemptionList {
  redemptions: Redemption[];
}