# lib/api-zod/src/generated/types/reviewRedemptionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewRedemptionBodyAction } from "./reviewRedemptionBodyAction";

export interface ReviewRedemptionBody {
  action: ReviewRedemptionBodyAction;
  /** @nullable */
  note?: string | null;
}